+++ src/features/game/types/game.ts
@@ -78,6 +78,7 @@ import { GameTransaction } from "./transactions";
 import { CompetitionName, CompetitionProgress } from "./competitions";
 import { AnimalType } from "./animals";
 import { ChoreBoard } from "./choreBoard";
+import { RecipeItemName, Recipes } from "../lib/crafting";
 
 export type Reward = {
   coins?: number;
@@ -1371,9 +1372,10 @@ export interface GameState {
 
   craftingBox: {
     status: "pending" | "idle" | "crafting";
-    item?: InventoryItemName;
+    item?: RecipeItemName;
     startedAt: number;
     readyAt: number;
+    recipes: Partial<Recipes>;
   };
 }
